refactor(ui-react): decompose DeviceDetails into focused sub-components
DeviceDetails.tsx was 816 lines, mixing device info display, inline tag
CRUD, inline rename, inline custom fields, and an inline delete modal
— all in one file.

Extract four sub-components:
- InfoItem: reusable <dt>/<dd> display row with optional copy button
- TagsSection: tag CRUD (max 3, alphanumeric, 3–255 chars), with
  aria-label on add/remove buttons
- RenameSection: inline rename with Enter=save / Escape=cancel
- CustomFieldsSection: key-value pairs with inline confirm-to-delete

Replace the inline delete modal with ConfirmDialog (with error state).
DeviceDetails drops from 816 → 444 lines.
